* compileable with C++98 * factored out RegexReplace into a separate file * error handling commented out -> need to find a way to pass the info
