+++ Source/CMakeLists.txt
@@ -45,8 +45,6 @@ endif()
 
 set(EXECUTABLE_OUTPUT_PATH ${CMake_BIN_DIR})
 
-add_definitions(-std=c++11)
-
 # ensure Unicode friendly APIs are used on Windows
 if(WIN32)
   add_definitions(-DUNICODE -D_UNICODE)
@@ -321,6 +319,8 @@ set(SRCS
   cmPropertyMap.h
   cmQtAutoGenerators.cxx
   cmQtAutoGenerators.h
+  cmRegexTools.h
+  cmRegexTools.cxx
   cmRST.cxx
   cmRST.h
   cmScriptGenerator.h
