+++ Source/cmRegexTools.h
@@ -1,0 +1,20 @@
+/*============================================================================
+  CMake - Cross Platform Makefile Generator
+  Copyright 2000-2009 Kitware, Inc., Insight Software Consortium
+
+  Distributed under the OSI-approved BSD License (the "License");
+  see accompanying file Copyright.txt for details.
+
+  This software is distributed WITHOUT ANY WARRANTY; without even the
+  implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+  See the License for more information.
+============================================================================*/
+#ifndef cmRegexTools_h
+#define cmRegexTools_h
+
+#include <string>
+
+std::string RegexReplace(std::string const& input, std::string const& regex, std::string const& replace);
+
+
+#endif
